Content is not formatted on lines that the user input.
Previously there was an issue where if you typed something like
echo \e[24mHello, it would format as blue both when it echoes *and* when
displaying the user input. This was unintended behavior and now any
lines in the history that represent what the user typed (that is, any
lines prefaced with a prompt like default.user@portfolio:) are left
unformatted.
